refactor: 폴더 구조 pages, components, auth 에 따라 구분하여 분리함
