+++ src/components/Header.js
@@ -1,0 +1,9 @@
+function Footer() {
+  return `
+        <footer class="bg-gray-200 p-4 text-center">
+          <p>&copy; 2024 항해플러스. All rights reserved.</p>
+        </footer>
+      `;
+}
+
+export default Footer;
